Move drop effect to graphic layer
